Split contexts to user and admin.
