+++ features/user_homepage.feature
@@ -7,4 +7,3 @@ Feature: See homepage
         Given I am on store homepage
          Then I should be on store homepage
           And I should see "Recent products"
-          And I should see 0 products on the page
